Tidy up View names
