+++ OnnxStack.UI/Views/ImageInpaintView.xaml.cs
@@ -1,5 +1,4 @@
-﻿using LibGit2Sharp;
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 using Models;
 using OnnxStack.StableDiffusion.Common;
 using OnnxStack.StableDiffusion.Config;
@@ -22,11 +21,11 @@ using System.Windows.Controls;
 namespace OnnxStack.UI.Views
 {
     /// <summary>
-    /// Interaction logic for ImageToImage.xaml
+    /// Interaction logic for ImageInpaintView.xaml
     /// </summary>
-    public partial class ImageToImage : UserControl, INavigatable, INotifyPropertyChanged
+    public partial class ImageInpaintView : UserControl, INavigatable, INotifyPropertyChanged
     {
-        private readonly ILogger<ImageToImage> _logger;
+        private readonly ILogger<ImageInpaintView> _logger;
         private readonly IStableDiffusionService _stableDiffusionService;
 
         private bool _hasResult;
@@ -35,9 +34,11 @@ namespace OnnxStack.UI.Views
         private bool _isGenerating;
         private int _selectedTabIndex;
         private bool _hasInputResult;
+        private bool _hasInputMaskResult;
         private bool _isControlsEnabled;
         private ImageInput _inputImage;
         private ImageResult _resultImage;
+        private ImageInput _inputImageMask;
         private ModelOptionsModel _selectedModel;
         private PromptOptionsModel _promptOptionsModel;
         private SchedulerOptionsModel _schedulerOptions;
@@ -46,22 +47,22 @@ namespace OnnxStack.UI.Views
 
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="ImageToImage"/> class.
+        /// Initializes a new instance of the <see cref="ImageInpaintView"/> class.
         /// </summary>
-        public ImageToImage()
+        public ImageInpaintView()
         {
             if (!DesignerProperties.GetIsInDesignMode(this))
             {
-                _logger = App.GetService<ILogger<ImageToImage>>();
+                _logger = App.GetService<ILogger<ImageInpaintView>>();
                 _stableDiffusionService = App.GetService<IStableDiffusionService>();
             }
 
-            SupportedDiffusers = new() { DiffuserType.ImageToImage };
+            SupportedDiffusers = new() { DiffuserType.ImageInpaint, DiffuserType.ImageInpaintLegacy };
             CancelCommand = new AsyncRelayCommand(Cancel, CanExecuteCancel);
             GenerateCommand = new AsyncRelayCommand(Generate, CanExecuteGenerate);
             ClearHistoryCommand = new AsyncRelayCommand(ClearHistory, CanExecuteClearHistory);
             PromptOptions = new PromptOptionsModel();
-            SchedulerOptions = new SchedulerOptionsModel();
+            SchedulerOptions = new SchedulerOptionsModel { SchedulerType = SchedulerType.DDPM };
             BatchOptions = new BatchOptionsModel();
             ImageResults = new ObservableCollection<ImageResult>();
             ProgressMax = SchedulerOptions.InferenceSteps;
@@ -75,7 +76,7 @@ namespace OnnxStack.UI.Views
             set { SetValue(UISettingsProperty, value); }
         }
         public static readonly DependencyProperty UISettingsProperty =
-            DependencyProperty.Register("UISettings", typeof(OnnxStackUIConfig), typeof(ImageToImage));
+            DependencyProperty.Register("UISettings", typeof(OnnxStackUIConfig), typeof(ImageInpaintView));
 
         public List<DiffuserType> SupportedDiffusers { get; }
         public AsyncRelayCommand CancelCommand { get; }
@@ -119,6 +120,12 @@ namespace OnnxStack.UI.Views
             set { _inputImage = value; NotifyPropertyChanged(); }
         }
 
+        public ImageInput InputImageMask
+        {
+            get { return _inputImageMask; }
+            set { _inputImageMask = value; NotifyPropertyChanged(); }
+        }
+
         public int ProgressValue
         {
             get { return _progressValue; }
@@ -149,6 +156,13 @@ namespace OnnxStack.UI.Views
             set { _hasInputResult = value; NotifyPropertyChanged(); }
         }
 
+        public bool HasInputMaskResult
+        {
+            get { return _hasInputMaskResult; }
+            set { _hasInputMaskResult = value; NotifyPropertyChanged(); }
+        }
+
+
         public int SelectedTabIndex
         {
             get { return _selectedTabIndex; }
@@ -172,8 +186,11 @@ namespace OnnxStack.UI.Views
             Reset();
             HasResult = false;
             ResultImage = null;
+            InputImageMask = null;
             HasInputResult = true;
-            if (imageResult.Model.ModelOptions.Diffusers.Contains(DiffuserType.ImageToImage))
+            HasInputMaskResult = false;
+            if (imageResult.Model.ModelOptions.Diffusers.Contains(DiffuserType.ImageInpaint)
+             || imageResult.Model.ModelOptions.Diffusers.Contains(DiffuserType.ImageInpaintLegacy))
             {
                 SelectedModel = imageResult.Model;
             }
@@ -202,9 +219,10 @@ namespace OnnxStack.UI.Views
             IsGenerating = true;
             IsControlsEnabled = false;
             ResultImage = null;
-            var promptOptions = GetPromptOptions(PromptOptions, InputImage);
+            var promptOptions = GetPromptOptions(PromptOptions, InputImage, InputImageMask);
             var batchOptions = BatchOptions.ToBatchOptions();
             var schedulerOptions = SchedulerOptions.ToSchedulerOptions();
+            schedulerOptions.Strength = 1; // Make sure strength is 1 for Image Inpainting
 
             try
             {
@@ -343,11 +361,12 @@ namespace OnnxStack.UI.Views
                 while (!_cancelationTokenSource.IsCancellationRequested)
                 {
                     var refreshTimestamp = Stopwatch.GetTimestamp();
-                    if (SchedulerOptions.HasChanged || PromptOptions.HasChanged)
+                    if (SchedulerOptions.HasChanged || PromptOptions.HasChanged || HasInputMaskResult)
                     {
+                        HasInputMaskResult = false;
                         PromptOptions.HasChanged = false;
                         SchedulerOptions.HasChanged = false;
-                        var realtimePromptOptions = GetPromptOptions(PromptOptions, InputImage);
+                        var realtimePromptOptions = GetPromptOptions(PromptOptions, InputImage, InputImageMask);
                         var realtimeSchedulerOptions = SchedulerOptions.ToSchedulerOptions();
 
                         var timestamp = Stopwatch.GetTimestamp();
@@ -368,16 +387,22 @@ namespace OnnxStack.UI.Views
         }
 
 
-        private PromptOptions GetPromptOptions(PromptOptionsModel promptOptionsModel, ImageInput imageInput)
+        private PromptOptions GetPromptOptions(PromptOptionsModel promptOptionsModel, ImageInput imageInput, ImageInput imageInputMask)
         {
             return new PromptOptions
             {
                 Prompt = promptOptionsModel.Prompt,
                 NegativePrompt = promptOptionsModel.NegativePrompt,
-                DiffuserType = DiffuserType.ImageToImage,
+                DiffuserType = SelectedModel.ModelOptions.Diffusers.Contains(DiffuserType.ImageInpaint)
+                    ? DiffuserType.ImageInpaint
+                    : DiffuserType.ImageInpaintLegacy,
                 InputImage = new StableDiffusion.Models.InputImage
                 {
                     ImageBytes = imageInput.Image.GetImageBytes()
+                },
+                InputImageMask = new StableDiffusion.Models.InputImage
+                {
+                    ImageBytes = imageInputMask.Image.GetImageBytes()
                 }
             };
         }
@@ -409,7 +434,7 @@ namespace OnnxStack.UI.Views
             };
 
             if (UISettings.ImageAutoSave)
-                await imageResult.AutoSaveAsync(Path.Combine(UISettings.ImageAutoSaveDirectory, "ImageToImage"), UISettings.ImageAutoSaveBlueprint);
+                await imageResult.AutoSaveAsync(Path.Combine(UISettings.ImageAutoSaveDirectory, "ImageInpaint"), UISettings.ImageAutoSaveBlueprint);
             return imageResult;
         }
