+++ OnnxStack.UI/Views/ImageToImageView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="OnnxStack.UI.Views.ImageInpaint"
+﻿<UserControl x:Class="OnnxStack.UI.Views.ImageToImageView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -21,18 +21,18 @@
             </UniformGrid>
             <DockPanel>
                 <userControls:ModelPickerControl DockPanel.Dock="Top"
-                    UISettings="{Binding UISettings}"
-                    SupportedDiffusers="{Binding SupportedDiffusers}"
-                    IsEnabled="{Binding IsGenerating, Converter={StaticResource InverseBoolConverter}}"
-                    SelectedModel="{Binding SelectedModel, Mode=TwoWay}" Models="{Binding RelativeSource={RelativeSource FindAncestor,AncestorType={x:Type Window}}, Path=Models}"/>
+                      UISettings="{Binding UISettings}"
+                      SupportedDiffusers="{Binding SupportedDiffusers}"
+                      IsEnabled="{Binding IsGenerating, Converter={StaticResource InverseBoolConverter}}"
+                      SelectedModel="{Binding SelectedModel, Mode=TwoWay}" Models="{Binding RelativeSource={RelativeSource FindAncestor,AncestorType={x:Type Window}}, Path=Models}"/>
                 <UniformGrid Rows="2" IsEnabled="{Binding SelectedModel.IsLoaded, FallbackValue=False, TargetNullValue=False}">
-                    <userControls:PromptControl
+                    <userControls:PromptControl 
                         SelectedModel="{Binding SelectedModel}"
                         PromptOptions="{Binding PromptOptions}" 
                         IsEnabled="{Binding IsControlsEnabled}"/>
                     <userControls:SchedulerControl DockPanel.Dock="Bottom" Margin="0, 10, 0 ,0"
                         SelectedModel="{Binding SelectedModel}"
-                        DiffuserType="ImageInpaint" 
+                        DiffuserType="ImageToImage" 
                         BatchOptions="{Binding BatchOptions}"
                         SchedulerOptions="{Binding SchedulerOptions, Mode=TwoWay}" 
                         IsEnabled="{Binding IsControlsEnabled}"
@@ -60,10 +60,8 @@
                             <userControls:ImageInputControl 
                                 Result="{Binding InputImage, Mode=TwoWay}" 
                                 HasResult="{Binding HasInputResult, Mode=TwoWay}" 
-                                MaskResult="{Binding InputImageMask, Mode=TwoWay}"
-                                HasMaskResult="{Binding HasInputMaskResult, Mode=TwoWay}"
                                 SchedulerOptions="{Binding SchedulerOptions}"
-                                IsMaskEnabled="True"/>
+                                IsMaskEnabled="False"/>
                         </Viewbox>
                         <Viewbox Margin="20,40,40,40" >
                             <userControls:ImageResultControl
