+++ OnnxStack.UI/Views/LoggerView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="OnnxStack.UI.Views.Logger"
+﻿<UserControl x:Class="OnnxStack.UI.Views.LoggerView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
