+++ OnnxStack.UI/Views/SettingsView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="OnnxStack.UI.Views.Settings"
+﻿<UserControl x:Class="OnnxStack.UI.Views.SettingsView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
