+++ OnnxStack.UI/Views/TextToImageView.xaml
