+++ OnnxStack.UI/Views/TextToImageView.xaml.cs
